add force_disconnect function
Forces the client to disconnect from the tracker, causing it to reconnect
when the next request is made.  It will reconnect to a different tracker if
possible.  A paranoid application may wish to do to this before retrying a
failed command, on the off chance that another tracker may be working better.

[ew: commit message body taken from POD]

Signed-off-by: Eric Wong <normalperson@yhbt.net>